Move things around
This commit attempts to reconcile the code that belongs in entry, relay,
and relaychain. We find boundaries which work and don't cause import
cycles. Unfortuantely a large commit.
