+++ pkg/beacon/relay/chain/chain.go
@@ -3,11 +3,10 @@ package chain
 import (
 	"math/big"
 
-	"github.com/keep-network/keep-core/pkg/beacon/relay"
 	"github.com/keep-network/keep-core/pkg/beacon/relay/config"
 )
 
-// Interface represents the interface that the relay expects to interact
+// ChainInterface represents the interface that the relay expects to interact
 // with the anchoring blockchain on.
 type Interface interface {
 	// GetConfig returns the expected configuration of the threshold relay.
@@ -30,9 +29,4 @@ type Interface interface {
 	//
 	// TODO activation delay may be unnecessary, we'll see.
 	OnGroupPublicKeySubmitted(func(groupID string, activationBlock *big.Int)) error
-	// SubmitRelayEntryCandidate takes an entry (a group signature and
-	// timestamp), constructs a relay response from the elements,
-	// and attempts to submit the relay reponse to the chain for consideration.
-	// If the chain accepts the response as valid...
-	SubmitRelayEntryCandidate(relay.Entry) error
 }
